+++ lib/regex/character.rb
@@ -25,7 +25,8 @@ module Regex # This module is used as a namespace
       "\\7" => 7  
     }.freeze
 
-    MetaChars = '\^$+?.'.freeze
+    MetaChars = '\^$.|+?*()[]{}'.freeze
+    MetaCharsInClass = '\^[]-'.freeze # Characters with special meaning in char. class
 
     # The integer value that uniquely identifies the character.
     attr_reader(:codepoint)